fix: quilc startup command and refactor provider to setup.py
