+++ tests/runtime/rigetti/test_provider.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# pylint: disable=redefined-outer-name,too-many-lines,import-outside-toplevel,possibly-used-before-assignment
+# pylint: disable=redefined-outer-name,import-outside-toplevel,possibly-used-before-assignment
 
 """Unit tests for RigettiProvider."""
 
@@ -33,6 +33,7 @@ pytestmark = pytest.mark.skipif(not pyquil_found, reason="pyquil not installed")
 
 if pyquil_found:
     from qbraid.runtime.rigetti import RigettiDevice, RigettiProvider
+    from qbraid.runtime.rigetti.setup import RigettiProviderError
 else:
     RigettiDevice = None
     RigettiProvider = None
@@ -75,10 +76,10 @@ class TestRigettiProviderInit:
         monkeypatch.setenv("RIGETTI_ISSUER", DUMMY_ISSUER)
 
         with (
-            patch("qbraid.runtime.rigetti.provider.AuthServer") as mock_auth_server_cls,
-            patch("qbraid.runtime.rigetti.provider.OAuthSession"),
-            patch("qbraid.runtime.rigetti.provider.RefreshToken"),
-            patch("qbraid.runtime.rigetti.provider.QCSClient") as mock_qcs_client_cls,
+            patch(f"{_SETUP_MODULE}.AuthServer") as mock_auth_server_cls,
+            patch(f"{_SETUP_MODULE}.OAuthSession"),
+            patch(f"{_SETUP_MODULE}.RefreshToken"),
+            patch(f"{_SETUP_MODULE}.QCSClient") as mock_qcs_client_cls,
             patch.object(RigettiProvider, "_build_execution_options", return_value=MagicMock()),
         ):
             mock_auth_server_cls.return_value = MagicMock(name="CustomAuthServer")
@@ -101,10 +102,10 @@ class TestRigettiProviderInit:
         monkeypatch.setenv("RIGETTI_ISSUER", DUMMY_ISSUER)
 
         with (
-            patch("qbraid.runtime.rigetti.provider.AuthServer") as mock_auth_server_cls,
-            patch("qbraid.runtime.rigetti.provider.OAuthSession"),
-            patch("qbraid.runtime.rigetti.provider.RefreshToken"),
-            patch("qbraid.runtime.rigetti.provider.QCSClient") as mock_qcs_client_cls,
+            patch(f"{_SETUP_MODULE}.AuthServer") as mock_auth_server_cls,
+            patch(f"{_SETUP_MODULE}.OAuthSession"),
+            patch(f"{_SETUP_MODULE}.RefreshToken"),
+            patch(f"{_SETUP_MODULE}.QCSClient") as mock_qcs_client_cls,
             patch.object(RigettiProvider, "_build_execution_options", return_value=MagicMock()),
         ):
             mock_auth_server_cls.default.return_value = MagicMock(name="DefaultAuthServer")
@@ -124,10 +125,10 @@ class TestRigettiProviderInit:
         monkeypatch.delenv("RIGETTI_ISSUER", raising=False)
 
         with (
-            patch("qbraid.runtime.rigetti.provider.AuthServer") as mock_auth_server_cls,
-            patch("qbraid.runtime.rigetti.provider.OAuthSession"),
-            patch("qbraid.runtime.rigetti.provider.RefreshToken"),
-            patch("qbraid.runtime.rigetti.provider.QCSClient") as mock_qcs_client_cls,
+            patch(f"{_SETUP_MODULE}.AuthServer") as mock_auth_server_cls,
+            patch(f"{_SETUP_MODULE}.OAuthSession"),
+            patch(f"{_SETUP_MODULE}.RefreshToken"),
+            patch(f"{_SETUP_MODULE}.QCSClient") as mock_qcs_client_cls,
             patch.object(RigettiProvider, "_build_execution_options", return_value=MagicMock()),
         ):
             mock_auth_server_cls.default.return_value = MagicMock(name="DefaultAuthServer")
@@ -147,10 +148,10 @@ class TestRigettiProviderInit:
         monkeypatch.delenv("RIGETTI_ISSUER", raising=False)
 
         with (
-            patch("qbraid.runtime.rigetti.provider.AuthServer") as mock_auth_server_cls,
-            patch("qbraid.runtime.rigetti.provider.OAuthSession"),
-            patch("qbraid.runtime.rigetti.provider.RefreshToken"),
-            patch("qbraid.runtime.rigetti.provider.QCSClient") as mock_qcs_client_cls,
+            patch(f"{_SETUP_MODULE}.AuthServer") as mock_auth_server_cls,
+            patch(f"{_SETUP_MODULE}.OAuthSession"),
+            patch(f"{_SETUP_MODULE}.RefreshToken"),
+            patch(f"{_SETUP_MODULE}.QCSClient") as mock_qcs_client_cls,
             patch.object(RigettiProvider, "_build_execution_options", return_value=MagicMock()),
         ):
             mock_auth_server_cls.default.return_value = MagicMock(name="DefaultAuthServer")
@@ -170,10 +171,10 @@ class TestRigettiProviderInit:
         monkeypatch.delenv("RIGETTI_ISSUER", raising=False)
 
         with (
-            patch("qbraid.runtime.rigetti.provider.AuthServer") as mock_auth_server_cls,
-            patch("qbraid.runtime.rigetti.provider.OAuthSession") as mock_oauth_session_cls,
-            patch("qbraid.runtime.rigetti.provider.RefreshToken") as mock_refresh_token_cls,
-            patch("qbraid.runtime.rigetti.provider.QCSClient") as mock_qcs_client_cls,
+            patch(f"{_SETUP_MODULE}.AuthServer") as mock_auth_server_cls,
+            patch(f"{_SETUP_MODULE}.OAuthSession") as mock_oauth_session_cls,
+            patch(f"{_SETUP_MODULE}.RefreshToken") as mock_refresh_token_cls,
+            patch(f"{_SETUP_MODULE}.QCSClient") as mock_qcs_client_cls,
             patch.object(RigettiProvider, "_build_execution_options", return_value=MagicMock()),
         ):
             fake_auth_server = MagicMock(name="DefaultAuthServer")
@@ -442,270 +443,8 @@ class TestRigettiProviderGetDevice:
 if pyquil_found:
     from subprocess import TimeoutExpired
 
-    from qbraid.runtime.rigetti.provider import RigettiProviderError
-
 _PROVIDER_MODULE = "qbraid.runtime.rigetti.provider"
-
-
-class TestBuildQcsClient:
-    """Tests for the static method RigettiProvider._build_qcs_client."""
-
-    def test_build_qcs_client_uses_default_auth_server_when_no_client_id(self) -> None:
-        """When client_id is None, AuthServer.default() should be used."""
-        with (
-            patch(f"{_PROVIDER_MODULE}.AuthServer") as mock_auth_cls,
-            patch(f"{_PROVIDER_MODULE}.OAuthSession") as mock_oauth_cls,
-            patch(f"{_PROVIDER_MODULE}.RefreshToken") as mock_rt_cls,
-            patch(f"{_PROVIDER_MODULE}.QCSClient") as mock_qcs_cls,
-        ):
-            fake_auth = MagicMock(name="DefaultAuthServer")
-            mock_auth_cls.default.return_value = fake_auth
-            mock_rt_cls.return_value = MagicMock(name="RefreshToken")
-            mock_oauth_cls.return_value = MagicMock(name="OAuthSession")
-            mock_qcs_cls.return_value = MagicMock(name="QCSClient")
-
-            RigettiProvider._build_qcs_client(refresh_token=DUMMY_TOKEN)
-
-            mock_auth_cls.default.assert_called_once()
-            mock_auth_cls.assert_not_called()
-
-    def test_build_qcs_client_uses_default_auth_server_when_no_issuer(self) -> None:
-        """When issuer is None, AuthServer.default() should be used."""
-        with (
-            patch(f"{_PROVIDER_MODULE}.AuthServer") as mock_auth_cls,
-            patch(f"{_PROVIDER_MODULE}.OAuthSession"),
-            patch(f"{_PROVIDER_MODULE}.RefreshToken"),
-            patch(f"{_PROVIDER_MODULE}.QCSClient"),
-        ):
-            mock_auth_cls.default.return_value = MagicMock()
-
-            RigettiProvider._build_qcs_client(
-                refresh_token=DUMMY_TOKEN,
-                client_id=DUMMY_CLIENT_ID,
-                issuer=None,
-            )
-
-            mock_auth_cls.default.assert_called_once()
-            mock_auth_cls.assert_not_called()
-
-    def test_build_qcs_client_uses_custom_auth_server_when_both_provided(self) -> None:
-        """When both client_id and issuer are given, AuthServer(client_id, issuer) is used."""
-        with (
-            patch(f"{_PROVIDER_MODULE}.AuthServer") as mock_auth_cls,
-            patch(f"{_PROVIDER_MODULE}.OAuthSession"),
-            patch(f"{_PROVIDER_MODULE}.RefreshToken"),
-            patch(f"{_PROVIDER_MODULE}.QCSClient"),
-        ):
-            mock_auth_cls.return_value = MagicMock(name="CustomAuth")
-
-            RigettiProvider._build_qcs_client(
-                refresh_token=DUMMY_TOKEN,
-                client_id=DUMMY_CLIENT_ID,
-                issuer=DUMMY_ISSUER,
-            )
-
-            mock_auth_cls.assert_called_once_with(client_id=DUMMY_CLIENT_ID, issuer=DUMMY_ISSUER)
-            mock_auth_cls.default.assert_not_called()
-
-    def test_build_qcs_client_passes_url_kwargs(self) -> None:
-        """URL parameters should be forwarded to QCSClient constructor."""
-        with (
-            patch(f"{_PROVIDER_MODULE}.AuthServer") as mock_auth_cls,
-            patch(f"{_PROVIDER_MODULE}.OAuthSession") as mock_oauth_cls,
-            patch(f"{_PROVIDER_MODULE}.RefreshToken"),
-            patch(f"{_PROVIDER_MODULE}.QCSClient") as mock_qcs_cls,
-        ):
-            mock_auth_cls.default.return_value = MagicMock()
-            fake_session = MagicMock(name="OAuthSession")
-            mock_oauth_cls.return_value = fake_session
-
-            grpc = "https://custom-grpc:443"
-            quilc = "tcp://custom-quilc:5555"
-            qvm = "http://custom-qvm:5000"
-
-            RigettiProvider._build_qcs_client(
-                refresh_token=DUMMY_TOKEN,
-                grpc_api_url=grpc,
-                quilc_url=quilc,
-                qvm_url=qvm,
-            )
-
-            mock_qcs_cls.assert_called_once_with(
-                oauth_session=fake_session,
-                grpc_api_url=grpc,
-                quilc_url=quilc,
-                qvm_url=qvm,
-            )
-
-    def test_build_qcs_client_omits_url_kwargs_when_none(self) -> None:
-        """When URL params are None they should not appear in QCSClient kwargs."""
-        with (
-            patch(f"{_PROVIDER_MODULE}.AuthServer") as mock_auth_cls,
-            patch(f"{_PROVIDER_MODULE}.OAuthSession") as mock_oauth_cls,
-            patch(f"{_PROVIDER_MODULE}.RefreshToken"),
-            patch(f"{_PROVIDER_MODULE}.QCSClient") as mock_qcs_cls,
-        ):
-            mock_auth_cls.default.return_value = MagicMock()
-            fake_session = MagicMock(name="OAuthSession")
-            mock_oauth_cls.return_value = fake_session
-
-            RigettiProvider._build_qcs_client(
-                refresh_token=DUMMY_TOKEN,
-                grpc_api_url=None,
-                quilc_url=None,
-                qvm_url=None,
-            )
-
-            mock_qcs_cls.assert_called_once_with(oauth_session=fake_session)
-
-
-class TestIsPortInUse:
-    """Tests for the static method RigettiProvider._is_port_in_use."""
-
-    def test_returns_true_when_port_is_open(self) -> None:
-        """A successful connect_ex (returns 0) means the port is in use."""
-        mock_sock = MagicMock()
-        mock_sock.connect_ex.return_value = 0
-        mock_sock.__enter__ = MagicMock(return_value=mock_sock)
-        mock_sock.__exit__ = MagicMock(return_value=False)
-
-        with patch("qbraid.runtime.rigetti.provider.socket.socket", return_value=mock_sock):
-            assert RigettiProvider._is_port_in_use(5555) is True
-
-        mock_sock.connect_ex.assert_called_once_with(("127.0.0.1", 5555))
-
-    def test_returns_false_when_port_is_closed(self) -> None:
-        """A non-zero connect_ex means the port is not in use."""
-        mock_sock = MagicMock()
-        mock_sock.connect_ex.return_value = 111  # ECONNREFUSED
-        mock_sock.__enter__ = MagicMock(return_value=mock_sock)
-        mock_sock.__exit__ = MagicMock(return_value=False)
-
-        with patch("qbraid.runtime.rigetti.provider.socket.socket", return_value=mock_sock):
-            assert RigettiProvider._is_port_in_use(5555) is False
-
-    def test_uses_custom_host(self) -> None:
-        """The host parameter should be forwarded to connect_ex."""
-        mock_sock = MagicMock()
-        mock_sock.connect_ex.return_value = 0
-        mock_sock.__enter__ = MagicMock(return_value=mock_sock)
-        mock_sock.__exit__ = MagicMock(return_value=False)
-
-        with patch("qbraid.runtime.rigetti.provider.socket.socket", return_value=mock_sock):
-            RigettiProvider._is_port_in_use(9999, host="10.0.0.1")
-
-        mock_sock.connect_ex.assert_called_once_with(("10.0.0.1", 9999))
-
-
-class TestFindBinary:
-    """Tests for the static method RigettiProvider._find_binary."""
-
-    def test_returns_path_from_shutil_which(self) -> None:
-        """When shutil.which finds the binary, its path is returned."""
-        with patch("qbraid.runtime.rigetti.provider.shutil.which", return_value="/usr/bin/quilc"):
-            result = RigettiProvider._find_binary("quilc")
-
-        from pathlib import Path
-
-        assert result == Path("/usr/bin/quilc")
-
-    def test_falls_back_to_qbraid_bin_dir(self) -> None:
-        """When shutil.which returns None, check ~/.qbraid/rigetti/bin/."""
-        from pathlib import Path
-
-        fallback_path = Path.home() / ".qbraid" / "rigetti" / "bin" / "quilc"
-
-        with (
-            patch("qbraid.runtime.rigetti.provider.shutil.which", return_value=None),
-            patch.object(Path, "is_file", return_value=True),
-            patch("qbraid.runtime.rigetti.provider.os.access", return_value=True),
-        ):
-            result = RigettiProvider._find_binary("quilc")
-
-        assert result == fallback_path
-
-    def test_returns_none_when_not_found(self) -> None:
-        """When neither shutil.which nor the fallback finds the binary, return None."""
-        from pathlib import Path
-
-        with (
-            patch("qbraid.runtime.rigetti.provider.shutil.which", return_value=None),
-            patch.object(Path, "is_file", return_value=False),
-        ):
-            result = RigettiProvider._find_binary("quilc")
-
-        assert result is None
-
-    def test_returns_none_when_fallback_not_executable(self) -> None:
-        """When the fallback file exists but is not executable, return None."""
-        from pathlib import Path
-
-        with (
-            patch("qbraid.runtime.rigetti.provider.shutil.which", return_value=None),
-            patch.object(Path, "is_file", return_value=True),
-            patch("qbraid.runtime.rigetti.provider.os.access", return_value=False),
-        ):
-            result = RigettiProvider._find_binary("quilc")
-
-        assert result is None
-
-
-class TestWaitForPort:
-    """Tests for the static method RigettiProvider._wait_for_port."""
-
-    def test_returns_immediately_when_port_is_open(self) -> None:
-        """If the port is already accepting connections, return immediately."""
-        mock_sock = MagicMock()
-        mock_sock.connect_ex.return_value = 0
-        mock_sock.__enter__ = MagicMock(return_value=mock_sock)
-        mock_sock.__exit__ = MagicMock(return_value=False)
-
-        with (
-            patch("qbraid.runtime.rigetti.provider.socket.socket", return_value=mock_sock),
-            patch("qbraid.runtime.rigetti.provider.time.monotonic", side_effect=[0.0, 0.1]),
-        ):
-            # Should not raise
-            RigettiProvider._wait_for_port(5555, timeout=5.0)
-
-    def test_raises_on_timeout(self) -> None:
-        """If the port never opens, raise RigettiProviderError."""
-        mock_sock = MagicMock()
-        mock_sock.connect_ex.return_value = 111  # always refused
-        mock_sock.__enter__ = MagicMock(return_value=mock_sock)
-        mock_sock.__exit__ = MagicMock(return_value=False)
-
-        # Simulate time progressing past the deadline
-        with (
-            patch("qbraid.runtime.rigetti.provider.socket.socket", return_value=mock_sock),
-            patch(
-                "qbraid.runtime.rigetti.provider.time.monotonic",
-                side_effect=[0.0, 0.5, 100.0],
-            ),
-            patch("qbraid.runtime.rigetti.provider.time.sleep"),
-            pytest.raises(RigettiProviderError, match="Timed out waiting for port"),
-        ):
-            RigettiProvider._wait_for_port(5555, timeout=5.0)
-
-    def test_retries_until_port_opens(self) -> None:
-        """Should retry polling until the port starts accepting connections."""
-        mock_sock = MagicMock()
-        # First two checks fail, third succeeds
-        mock_sock.connect_ex.side_effect = [111, 111, 0]
-        mock_sock.__enter__ = MagicMock(return_value=mock_sock)
-        mock_sock.__exit__ = MagicMock(return_value=False)
-
-        with (
-            patch("qbraid.runtime.rigetti.provider.socket.socket", return_value=mock_sock),
-            patch(
-                "qbraid.runtime.rigetti.provider.time.monotonic",
-                side_effect=[0.0, 0.5, 1.0, 1.5, 2.0],
-            ),
-            patch("qbraid.runtime.rigetti.provider.time.sleep") as mock_sleep,
-        ):
-            RigettiProvider._wait_for_port(5555, timeout=10.0)
-
-        assert mock_sock.connect_ex.call_count == 3
-        assert mock_sleep.call_count == 2
+_SETUP_MODULE = "qbraid.runtime.rigetti.setup"
 
 
 class TestStartQuilc:
@@ -721,13 +460,13 @@ class TestStartQuilc:
 
         with (
             patch(f"{_PROVIDER_MODULE}.Popen", return_value=fake_proc) as mock_popen,
-            patch.object(provider, "_wait_for_port") as mock_wait,
+            patch(f"{_PROVIDER_MODULE}.wait_for_port") as mock_wait,
         ):
             provider._start_quilc(binary)
 
         mock_popen.assert_called_once()
         call_args = mock_popen.call_args
-        assert call_args[0][0] == [str(binary), "-S"]
+        assert call_args[0][0] == [str(binary), "-P", "-S", "-p", "5555"]
         mock_wait.assert_called_once_with(5555)
         assert provider._quilc_process is fake_proc
 
@@ -745,7 +484,7 @@ class TestStartQvm:
 
         with (
             patch(f"{_PROVIDER_MODULE}.Popen", return_value=fake_proc) as mock_popen,
-            patch.object(provider, "_wait_for_port") as mock_wait,
+            patch(f"{_PROVIDER_MODULE}.wait_for_port") as mock_wait,
         ):
             provider._start_qvm(binary)
 
@@ -892,15 +631,6 @@ class TestRegisterCleanup:
         mock_atexit.assert_called_once()
 
 
-class TestDownloadForestSdk:
-    """Tests for the static method RigettiProvider._download_forest_sdk."""
-
-    def test_download_forest_sdk_raises_rigetti_provider_error(self) -> None:
-        """Should raise RigettiProviderError with installation instructions."""
-        with pytest.raises(RigettiProviderError, match="quilc binary not found"):
-            RigettiProvider._download_forest_sdk()
-
-
 class TestSetup:
     """Tests for RigettiProvider.setup — the main orchestration method."""
 
@@ -912,12 +642,11 @@ class TestSetup:
         fake_new_client = MagicMock(name="NewQCSClient")
 
         with (
-            patch.object(
-                RigettiProvider,
-                "_build_qcs_client",
+            patch(
+                f"{_PROVIDER_MODULE}.build_qcs_client",
                 return_value=fake_new_client,
             ) as mock_build,
-            patch.object(RigettiProvider, "_is_port_in_use", return_value=True),
+            patch(f"{_PROVIDER_MODULE}.is_port_in_use", return_value=True),
             patch.object(provider, "_build_execution_options", return_value=MagicMock()),
             patch.object(provider, "_register_cleanup"),
         ):
@@ -944,12 +673,11 @@ class TestSetup:
 
         with (
             patch("builtins.input", side_effect=["my-token", "", ""]) as mock_input,
-            patch.object(
-                RigettiProvider,
-                "_build_qcs_client",
+            patch(
+                f"{_PROVIDER_MODULE}.build_qcs_client",
                 return_value=MagicMock(),
             ),
-            patch.object(RigettiProvider, "_is_port_in_use", return_value=True),
+            patch(f"{_PROVIDER_MODULE}.is_port_in_use", return_value=True),
             patch.object(provider, "_build_execution_options", return_value=MagicMock()),
             patch.object(provider, "_register_cleanup"),
         ):
@@ -975,12 +703,11 @@ class TestSetup:
         provider = RigettiProvider(qcs_client=mock_qcs_client)
 
         with (
-            patch.object(
-                RigettiProvider,
-                "_build_qcs_client",
+            patch(
+                f"{_PROVIDER_MODULE}.build_qcs_client",
                 return_value=MagicMock(),
             ),
-            patch.object(RigettiProvider, "_is_port_in_use") as mock_port,
+            patch(f"{_PROVIDER_MODULE}.is_port_in_use") as mock_port,
             patch.object(provider, "_start_quilc") as mock_start,
             patch.object(provider, "_build_execution_options", return_value=MagicMock()),
             patch.object(provider, "_register_cleanup"),
@@ -1001,12 +728,11 @@ class TestSetup:
         provider = RigettiProvider(qcs_client=mock_qcs_client)
 
         with (
-            patch.object(
-                RigettiProvider,
-                "_build_qcs_client",
+            patch(
+                f"{_PROVIDER_MODULE}.build_qcs_client",
                 return_value=MagicMock(),
             ),
-            patch.object(RigettiProvider, "_is_port_in_use", return_value=True),
+            patch(f"{_PROVIDER_MODULE}.is_port_in_use", return_value=True),
             patch.object(provider, "_start_quilc") as mock_start,
             patch.object(provider, "_build_execution_options", return_value=MagicMock()),
             patch.object(provider, "_register_cleanup"),
@@ -1030,13 +756,12 @@ class TestSetup:
         fake_binary = Path("/usr/local/bin/quilc")
 
         with (
-            patch.object(
-                RigettiProvider,
-                "_build_qcs_client",
+            patch(
+                f"{_PROVIDER_MODULE}.build_qcs_client",
                 return_value=MagicMock(),
             ),
-            patch.object(RigettiProvider, "_is_port_in_use", return_value=False),
-            patch.object(RigettiProvider, "_find_binary", return_value=fake_binary),
+            patch(f"{_PROVIDER_MODULE}.is_port_in_use", return_value=False),
+            patch(f"{_PROVIDER_MODULE}.find_binary", return_value=fake_binary),
             patch.object(provider, "_start_quilc") as mock_start,
             patch.object(provider, "_build_execution_options", return_value=MagicMock()),
             patch.object(provider, "_register_cleanup"),
@@ -1053,20 +778,18 @@ class TestSetup:
     def test_setup_calls_download_forest_sdk_when_binary_missing(
         self, mock_qcs_client: MagicMock
     ) -> None:
-        """When quilc binary is not found, _download_forest_sdk should be called."""
+        """When quilc binary is not found, download_forest_sdk should be called."""
         provider = RigettiProvider(qcs_client=mock_qcs_client)
 
         with (
-            patch.object(
-                RigettiProvider,
-                "_build_qcs_client",
+            patch(
+                f"{_PROVIDER_MODULE}.build_qcs_client",
                 return_value=MagicMock(),
             ),
-            patch.object(RigettiProvider, "_is_port_in_use", return_value=False),
-            patch.object(RigettiProvider, "_find_binary", return_value=None),
-            patch.object(
-                RigettiProvider,
-                "_download_forest_sdk",
+            patch(f"{_PROVIDER_MODULE}.is_port_in_use", return_value=False),
+            patch(f"{_PROVIDER_MODULE}.find_binary", return_value=None),
+            patch(
+                f"{_PROVIDER_MODULE}.download_forest_sdk",
                 side_effect=RigettiProviderError("not found"),
             ) as mock_download,
             patch.object(provider, "_build_execution_options", return_value=MagicMock()),
@@ -1087,9 +810,8 @@ class TestSetup:
         provider = RigettiProvider(qcs_client=mock_qcs_client)
 
         with (
-            patch.object(
-                RigettiProvider,
-                "_build_qcs_client",
+            patch(
+                f"{_PROVIDER_MODULE}.build_qcs_client",
                 return_value=MagicMock(),
             ),
             patch.object(provider, "_start_qvm") as mock_start,
@@ -1116,13 +838,12 @@ class TestSetup:
         fake_binary = Path("/usr/local/bin/qvm")
 
         with (
-            patch.object(
-                RigettiProvider,
-                "_build_qcs_client",
+            patch(
+                f"{_PROVIDER_MODULE}.build_qcs_client",
                 return_value=MagicMock(),
             ),
-            patch.object(RigettiProvider, "_is_port_in_use", return_value=False),
-            patch.object(RigettiProvider, "_find_binary", return_value=fake_binary),
+            patch(f"{_PROVIDER_MODULE}.is_port_in_use", return_value=False),
+            patch(f"{_PROVIDER_MODULE}.find_binary", return_value=fake_binary),
             patch.object(provider, "_start_qvm") as mock_start,
             patch.object(provider, "_build_execution_options", return_value=MagicMock()),
             patch.object(provider, "_register_cleanup"),
@@ -1141,9 +862,8 @@ class TestSetup:
         provider = RigettiProvider(qcs_client=mock_qcs_client)
 
         with (
-            patch.object(
-                RigettiProvider,
-                "_build_qcs_client",
+            patch(
+                f"{_PROVIDER_MODULE}.build_qcs_client",
                 return_value=MagicMock(),
             ),
             patch.object(provider, "_build_execution_options", return_value=MagicMock()),
@@ -1172,9 +892,8 @@ class TestSetup:
                 "builtins.input",
                 side_effect=[DUMMY_TOKEN, DUMMY_CLIENT_ID, DUMMY_ISSUER],
             ) as mock_input,
-            patch.object(
-                RigettiProvider,
-                "_build_qcs_client",
+            patch(
+                f"{_PROVIDER_MODULE}.build_qcs_client",
                 return_value=MagicMock(),
             ),
             patch.object(provider, "_build_execution_options", return_value=MagicMock()),
